+++ examples/framework-nestjs/src/modules/common/inngest/types/events.type.ts
@@ -1,0 +1,5 @@
+import { Hello } from '@modules/common/inngest/types';
+
+export type Events = {
+  'job/hello.world': Hello;
+};