+++ examples/framework-nextjs-app-router/src/app/api/inngest/route.ts
@@ -1,0 +1,7 @@
+import { functions, inngest } from "@/inngest";
+import { serve } from "inngest/next";
+
+export const { GET, POST, PUT } = serve({
+  client: inngest,
+  functions,
+});