+++ packages/inngest/src/components/Inngest.ts
@@ -2,7 +2,7 @@ import { type Exact } from "type-fest";
 import { InngestApi } from "../api/api";
 import {
   defaultDevServerHost,
-  defaultInngestBaseUrl,
+  defaultInngestApiBaseUrl,
   defaultInngestEventBaseUrl,
   dummyEventKey,
   envKeys,
@@ -103,7 +103,8 @@ export class Inngest<TOpts extends ClientOptions = ClientOptions> {
    */
   private eventKey = "";
 
-  private readonly baseUrl: string | undefined;
+  private readonly apiBaseUrl: string | undefined;
+  private readonly eventBaseUrl: string | undefined;
 
   private readonly inngestApi: InngestApi;
 
@@ -165,24 +166,17 @@ export class Inngest<TOpts extends ClientOptions = ClientOptions> {
 
     this.id = id;
 
-    this.baseUrl = baseUrl || processEnv(envKeys.InngestBaseUrl);
+    this.apiBaseUrl =
+      baseUrl ||
+      processEnv(envKeys.InngestApiBaseUrl) ||
+      processEnv(envKeys.InngestBaseUrl);
+
+    this.eventBaseUrl =
+      baseUrl ||
+      processEnv(envKeys.InngestEventApiBaseUrl) ||
+      processEnv(envKeys.InngestBaseUrl);
 
     this.setEventKey(eventKey || processEnv(envKeys.InngestEventKey) || "");
-
-    if (!this.eventKeySet()) {
-      console.warn(
-        prettyError({
-          type: "warn",
-          whatHappened: "Could not find event key",
-          consequences:
-            "Sending events will throw in production unless an event key is added.",
-          toFixNow: fixEventKeyMissingSteps,
-          why: "We couldn't find an event key to use to send events to Inngest.",
-          otherwise:
-            "Create a new production event key at https://app.inngest.com/env/production/manage/keys.",
-        })
-      );
-    }
 
     this.headers = inngestHeaders({
       inngestEnv: env,
@@ -191,7 +185,7 @@ export class Inngest<TOpts extends ClientOptions = ClientOptions> {
     this.fetch = getFetch(fetch);
 
     this.inngestApi = new InngestApi({
-      baseUrl: this.baseUrl || defaultInngestBaseUrl,
+      baseUrl: this.apiBaseUrl || defaultInngestApiBaseUrl,
       signingKey: processEnv(envKeys.InngestSigningKey) || "",
       fetch: this.fetch,
     });
@@ -297,7 +291,7 @@ export class Inngest<TOpts extends ClientOptions = ClientOptions> {
 
     this.sendEventUrl = new URL(
       `e/${this.eventKey}`,
-      this.baseUrl || defaultInngestEventBaseUrl
+      this.eventBaseUrl || defaultInngestEventBaseUrl
     );
   }
 
@@ -416,7 +410,7 @@ export class Inngest<TOpts extends ClientOptions = ClientOptions> {
      * user has set this it means they have already chosen a URL to hit.
      */
     if (!skipDevServer()) {
-      if (!this.baseUrl) {
+      if (!this.eventBaseUrl) {
         const devAvailable = await devServerAvailable(
           defaultDevServerHost,
           this.fetch